Clean chain helpers index
