+++ framework/src/modules/chain/helpers/index.js
@@ -1,6 +1,5 @@
 /**
  * @namespace helpers
- * @property {module:helpers/cache} Cache
  * @property {module:helpers/config} Config
  * @property {module:helpers/diff} Diff
  * @property {module:helpers/ed} Ed
@@ -8,6 +7,5 @@
  * @property {module:helpers/jobs_queue} Jobs_queue
  * @property {module:helpers/slots} Slots
  * @property {module:helpers/sort_by} Sort_by
- * @property {module:helpers/ws_api} WS_API
  * @property {module:helpers/z_schema} Z_schema
  */
